Fixed bugs with tutorial by refactoring the Distance calculator. Fixed bugs with bounce/death on enemies, complete refactor. Tweaked the timing of the harder patterns and wasps to be about 200 meters farther in the progression. Fixed a bug where the user would not properly spend yarn on upgrades (would attempt to spend for the next upgrade). Integrated new thorn trap asset.

+++ ProjectL/Assets/Scripts/Controllers/DistanceController.cs
@@ -18,13 +18,13 @@ public class DistanceController
         }
     }
 
-    private int _distance;
-    public int Distance
+    private float _distance;
+    public float Distance
     {
         get { return _distance; }
     }
 
-    public void AddToDistance(int distance)
+    public void AddToDistance(float distance)
     {
         _distance += distance;
     }
